feat(odd-platform-ui): split data quality layout (#1590)
